ref: Moved all console custom functions to its own file
