+++ CMakeLists.txt
@@ -42,6 +42,8 @@ add_library(Module SHARED
         Environment/Libraries/Metatable.hpp
         Environment/Libraries/Cache.cpp
         Environment/Libraries/Cache.hpp
+        Environment/Libraries/Console.cpp
+        Environment/Libraries/Console.hpp
 )
 target_include_directories(Module PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}")
 target_include_directories(Module PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/Dependencies")
