+++ Environment/Libraries/Globals.cpp
@@ -30,56 +30,6 @@ namespace RbxStu {
         return 1;
     }
 
-    int rconsolecreate(lua_State *L) {
-        AllocConsole();
-        return 0;
-    }
-
-    int rconsoledestroy(lua_State *L) {
-        FreeConsole();
-        return 0;
-    }
-
-    int rconsolesettitle(lua_State *L) {
-        auto wndName = luaL_checkstring(L, 1);
-        SetWindowTextA(GetConsoleWindow(), wndName);
-        return 0;
-    }
-
-    int rconsoleprint(lua_State *L) {
-        auto argc = lua_gettop(L);
-        std::stringstream strStream;
-        strStream << "[rconsoleprint] ";
-        for (int i = 0; i <= argc - 1; i++) {
-            const char *lStr = luaL_tolstring(L, i + 1, nullptr);
-            strStream << lStr << " ";
-        }
-        Logger::GetSingleton()->PrintInformation(RbxStu::RobloxConsole, strStream.str());
-    }
-
-    int rconsolewarn(lua_State *L) {
-        auto argc = lua_gettop(L);
-        std::stringstream strStream;
-        strStream << "[rconsolewarn] ";
-        for (int i = 0; i <= argc - 1; i++) {
-            const char *lStr = luaL_tolstring(L, i + 1, nullptr);
-            strStream << lStr << " ";
-        }
-        Logger::GetSingleton()->PrintWarning(RbxStu::RobloxConsole, strStream.str());
-    }
-
-    int rconsoleerror(lua_State *L) {
-        auto argc = lua_gettop(L);
-        std::stringstream strStream;
-        strStream << "[rconsoleerror] ";
-        for (int i = 0; i <= argc - 1; i++) {
-            const char *lStr = luaL_tolstring(L, i + 1, nullptr);
-            strStream << lStr << " ";
-        }
-        Logger::GetSingleton()->PrintError(RbxStu::RobloxConsole, strStream.str());
-        luaL_error(L, strStream.str().c_str());
-    }
-
     int getreg(lua_State *L) {
         lua_pushvalue(L, LUA_REGISTRYINDEX);
         return 1;
@@ -523,21 +473,6 @@ luaL_Reg *Globals::GetLibraryFunctions() {
                                {"decompile", RbxStu::decompile},
 
                                {"gettenv", RbxStu::gettenv},
-
-                               {"rconsolecreate", RbxStu::rconsolecreate},
-                               {"rconsoledestroy", RbxStu::rconsoledestroy},
-                               {"rconsolesettitle", RbxStu::rconsolesettitle},
-                               {"rconsolename", RbxStu::rconsolesettitle},
-                               {"rconsoleprint", RbxStu::rconsoleprint},
-                               {"rconsolewarn", RbxStu::rconsolewarn},
-                               {"rconsoleerror", RbxStu::rconsoleerror},
-
-                               {"consolecreate", RbxStu::rconsolecreate},
-                               {"consoledestroy", RbxStu::rconsoledestroy},
-                               {"consolesettitle", RbxStu::rconsolesettitle},
-                               {"consoleprint", RbxStu::rconsoleprint},
-                               {"consolewarn", RbxStu::rconsolewarn},
-                               {"consoleerror", RbxStu::rconsoleerror},
 
                                {nullptr, nullptr}};
     return reg;
